refactor feature spec with support
