+++ spec/features/create_group_with_public_spec.rb
@@ -1,19 +1,20 @@
 require 'rails_helper'
 
 RSpec.feature "CreateGroupWithPublics", type: :feature do
+  include CheckProductOptionSupport
   let(:user) { create(:user) }
   let(:last_group) { Group.last }
   let(:product) { Group.last.products.first}
   let(:option) { Group.last.options.first}
 
   before do
-    create(:group_with_option_and_product, is_public: true, title: "Public_Menu")
+    create(:group_with_option_and_product, is_public: true, title: "團團圓圓")
     sign_in user
     visit public_groups_path
   end
 
   scenario "進到 public 頁面可以看到公用菜單" do
-    expect(page).to have_content("Public_Menu")
+    expect(page).to have_content("團團圓圓")
   end
 
   scenario "按開團會進入新增團購頁面" do
@@ -21,7 +22,7 @@ RSpec.feature "CreateGroupWithPublics", type: :feature do
       click_link("開團")
     end
     expect(page).to have_content("來開團囉")
-    expect(page).to have_field('group[title]', with: "Public_Menu")
+    expect(page).to have_field('group[title]', with: "團團圓圓")
   end
 
   scenario "按確定開團會寫進資料庫", js: true, slow: true do
@@ -32,21 +33,12 @@ RSpec.feature "CreateGroupWithPublics", type: :feature do
       click_button("確定開團")
     end
 
-    expect(last_group.products.count).to be(1)
-    expect(last_group.options.count).to be(1)
-    expect(product.name).to eq("紅茶")
-    expect(product.price).to be(20)
-    expect(option.name).to eq("加椰果")
-    expect(option.price).to be(10)
-    expect(last_group.title).to eq("Public_Menu")
+    check_db_title_correct
 
-    visit edit_group_path(id: last_group.id)
-    expect(page).to have_field('group[title]', with: "Public_Menu")
-    click_on("菜單")
-    expect(page).to have_field("group[products_attributes][0][name]", with: "紅茶")
-    expect(page).to have_field("group[products_attributes][0][price]", with: "20")
-    click_on("選項")
-    expect(page).to have_field("group[options_attributes][0][name]", with: "加椰果")
-    expect(page).to have_field("group[options_attributes][0][price]", with: "10")
+    check_view_title_correct
+
+    check_db_product_option_correct
+
+    check_view_field_correct
   end
 end
